Separating document processing in dfm-data.
